Refactor FirebaseAuth service
