+++ lib/infrastructure/di/locator.dart
@@ -5,7 +5,8 @@ import '../../application/app_state_controller.dart';
 import '../../application/search/search_controller.dart';
 import '../repositories/mock_search_repository.dart';
 
-import '../auth/auth_service.dart'; // contiene también FirebaseAuthService
+import '../auth/auth_service.dart';
+import '../auth/firebase_auth_service.dart';
 
 final locator = GetIt.instance;
 
@@ -17,7 +18,9 @@ void setupLocator() {
   locator.registerLazySingleton<SearchRepository>(() => MockSearchRepository());
 
   // Auth
-  locator.registerLazySingleton<AuthService>(() => FirebaseAuthService());
+  if (!locator.isRegistered<AuthService>()) {
+    locator.registerLazySingleton<AuthService>(() => FirebaseAuthService());
+  }
 
   // Controllers
   locator.registerFactory<SearchController>(() => SearchController(locator()));
